Test using ifs

+++ Jenkinsfile
@@ -43,14 +43,16 @@ pipeline {
             sshagent(['ssh-amazon']) {
                 sh '''
                 ssh -o "StrictHostKeyChecking no" ec2-user@ec2-52-49-48-142.eu-west-1.compute.amazonaws.com "
-                docker stop $(docker ps -a -q)
-                docker rm $(docker ps -a -q)
+                if [ "$(docker ps -q)" ]; then
+                    docker stop $(docker ps -a -q)
+                    docker rm $(docker ps -a -q)
+                fi
                 docker pull ghcr.io/2000ghz/hello-2048/hello-2048:1.0.${BUILD_NUMBER}
                 docker run -td --rm -p 80:80 ghcr.io/2000ghz/hello-2048/hello-2048:1.0.${BUILD_NUMBER}"
                 '''
             }
         }
-    }
+}
 
 }
 }
